Initial implementation of vmm epoll crate
The vmm-epoll crate aims to manage and poll IO event notifications for
registered file decriptors by using epoll.

The epoll API performs a similar task to poll(2): monitoring multiple
file descriptors to see if I/O is possible on any of them. For a
typical usage mode, multiple file descriptors are registered on a epoll
fd, and a working thread polls and invokes event handler for IO event
notifications from the epoll fd.

Several traits are defined to support the above usage case:
- EpollEventMgr: manages a epoll fd and a pool of event slots. One file
  descriptor/event handler could be registered onto each slot.
- EpollSlotAllocator: allocates slots from the epoll event manager.
- EpollSlotGroup: a group of slots allocated from the epoll event
  manager.

Signed-off-by: Liu Jiang <gerry@linux.alibaba.com>