add negative test for EventSet errors
The added test uses UnixStream to check that when an error is received
on a fd which is registered to the inner epoll loop, the loop will
continuously yield the event until the corresponding fd is removed. The
test also check that once the fd is removed, no more errors are
returned.

Signed-off-by: Andreea Florescu <fandree@amazon.com>